backend/requirements.txt
# Core MQTT and data processing
paho-mqtt>=1.6.0
python-dotenv>=1.0.0

# InfluxDB integration
influxdb-client>=1.38.0

# Notifications
python-telegram-bot>=20.0
aiosmtplib>=2.0.0

# Data processing and scientific calculations
numpy>=1.21.0
pandas>=1.3.0

# Logging and monitoring
structlog>=23.0.0
colorlog>=6.7.0

# Configuration and utilities
pydantic>=2.0.0
python-dateutil>=2.8.0
pytz>=2023.3

# Optional: For advanced analytics
scipy>=1.7.0 